fix: handle uint80 round IDs in Chainlink price feeds
Chainlink's AggregatorV3 returns roundId and answeredInRound as uint80
using composite format: (phaseId << 64 | aggregatorRoundId). This can
exceed u64::MAX when phaseId > 0.

Changes:
- Change round_id and answered_in_round from u64 to u128 in PriceData
- Update PriceOutput struct in CLI to match
- Add test for large round IDs that would overflow u64

+++ crates/ethcli/src/cli/chainlink.rs
@@ -180,7 +180,8 @@ pub struct PriceOutput {
     pub price_usd: Option<f64>,
     pub raw_answer: String,
     pub decimals: u8,
-    pub round_id: u64,
+    /// Round ID (uint80 composite: phaseId << 64 | aggregatorRoundId)
+    pub round_id: u128,
     pub updated_at: u64,
     pub feed_address: Option<String>,
     pub is_stale: bool,
